pulled repeated styling out of bubbles into separate styling sheet
